backend/requirements.txt
fastapi[all]
uvicorn[standard]
motor
python-dotenv
pydantic
